Crate apache_avro_test_helper

Modules

data

Provides a set of Avro schema examples that are used in the tests.

logger

Structs

TestError

A custom error type for tests.

Functions

init

Does nothing. Just loads the crate. Should be used in the integration tests, because they do not use [dev-dependencies] and do not auto-load this crate.

Type Aliases

TestResult